move the message crate to sync module (#426)
