+++ core/src/sync/message/blocktxn.rs
@@ -2,7 +2,7 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use crate::{Message, MsgId, RequestId};
+use crate::sync::message::{Message, MsgId, RequestId};
 use cfx_types::H256;
 use primitives::TransactionWithSignature;
 use rlp::{Decodable, DecoderError, Encodable, Rlp, RlpStream};
