+++ core/src/sync/message/getblockheaderchain.rs
@@ -2,52 +2,52 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use crate::{Message, MsgId, RequestId};
+use crate::sync::message::{Message, MsgId, RequestId};
 use cfx_types::H256;
 use rlp::{Decodable, DecoderError, Encodable, Rlp, RlpStream};
 use std::ops::{Deref, DerefMut};
 
-#[derive(Debug, PartialEq, Default)]
-pub struct GetBlockTxn {
+#[derive(Debug, PartialEq, Clone)]
+pub struct GetBlockHeaderChain {
     pub request_id: RequestId,
-    pub block_hash: H256,
-    pub indexes: Vec<usize>,
+    pub hash: H256,
+    pub max_blocks: u64,
 }
 
-impl Message for GetBlockTxn {
-    fn msg_id(&self) -> MsgId { MsgId::GET_BLOCK_TXN }
+impl Message for GetBlockHeaderChain {
+    fn msg_id(&self) -> MsgId { MsgId::GET_BLOCK_HEADER_CHAIN }
 }
 
-impl Deref for GetBlockTxn {
+impl Deref for GetBlockHeaderChain {
     type Target = RequestId;
 
     fn deref(&self) -> &Self::Target { &self.request_id }
 }
 
-impl DerefMut for GetBlockTxn {
+impl DerefMut for GetBlockHeaderChain {
     fn deref_mut(&mut self) -> &mut RequestId { &mut self.request_id }
 }
 
-impl Encodable for GetBlockTxn {
+impl Encodable for GetBlockHeaderChain {
     fn rlp_append(&self, stream: &mut RlpStream) {
         stream
             .begin_list(3)
             .append(&self.request_id)
-            .append(&self.block_hash)
-            .append_list(&self.indexes);
+            .append(&self.hash)
+            .append(&self.max_blocks);
     }
 }
 
-impl Decodable for GetBlockTxn {
+impl Decodable for GetBlockHeaderChain {
     fn decode(rlp: &Rlp) -> Result<Self, DecoderError> {
         if rlp.item_count()? != 3 {
             return Err(DecoderError::RlpIncorrectListLen);
         }
 
-        Ok(GetBlockTxn {
+        Ok(GetBlockHeaderChain {
             request_id: rlp.val_at(0)?,
-            block_hash: rlp.val_at(1)?,
-            indexes: rlp.list_at(2)?,
+            hash: rlp.val_at(1)?,
+            max_blocks: rlp.val_at(2)?,
         })
     }
 }
