+++ core/src/sync/message/getblockheaders.rs
@@ -2,47 +2,49 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use crate::{Message, MsgId, RequestId};
-use primitives::Block;
+use crate::sync::message::{Message, MsgId, RequestId};
+use cfx_types::H256;
 use rlp::{Decodable, DecoderError, Encodable, Rlp, RlpStream};
 use std::ops::{Deref, DerefMut};
 
-#[derive(Debug, PartialEq, Default)]
-pub struct GetBlockBodiesResponse {
+#[derive(Debug, PartialEq, Clone)]
+pub struct GetBlockHeaders {
     pub request_id: RequestId,
-    pub bodies: Vec<Block>,
+    pub hashes: Vec<H256>,
 }
 
-impl Message for GetBlockBodiesResponse {
-    fn msg_id(&self) -> MsgId { MsgId::GET_BLOCK_BODIES_RESPONSE }
-
-    fn is_size_sensitive(&self) -> bool { self.bodies.len() > 0 }
+impl Message for GetBlockHeaders {
+    fn msg_id(&self) -> MsgId { MsgId::GET_BLOCK_HEADERS }
 }
 
-impl Deref for GetBlockBodiesResponse {
+impl Deref for GetBlockHeaders {
     type Target = RequestId;
 
     fn deref(&self) -> &Self::Target { &self.request_id }
 }
 
-impl DerefMut for GetBlockBodiesResponse {
+impl DerefMut for GetBlockHeaders {
     fn deref_mut(&mut self) -> &mut RequestId { &mut self.request_id }
 }
 
-impl Encodable for GetBlockBodiesResponse {
+impl Encodable for GetBlockHeaders {
     fn rlp_append(&self, stream: &mut RlpStream) {
         stream
             .begin_list(2)
             .append(&self.request_id)
-            .append_list(&self.bodies);
+            .append_list(&self.hashes);
     }
 }
 
-impl Decodable for GetBlockBodiesResponse {
+impl Decodable for GetBlockHeaders {
     fn decode(rlp: &Rlp) -> Result<Self, DecoderError> {
-        Ok(GetBlockBodiesResponse {
+        if rlp.item_count()? != 2 {
+            return Err(DecoderError::RlpIncorrectListLen);
+        }
+
+        Ok(GetBlockHeaders {
             request_id: rlp.val_at(0)?,
-            bodies: rlp.list_at(1)?,
+            hashes: rlp.list_at(1)?,
         })
     }
 }
