+++ core/src/sync/request_manager/mod.rs
@@ -4,13 +4,14 @@ use super::{
     },
     synchronization_state::SynchronizationState,
 };
-use crate::sync::Error;
-use cfx_types::H256;
-use message::{
-    GetBlockHashesByEpoch, GetBlockHeaderChain, GetBlockHeaders, GetBlockTxn,
-    GetBlocks, GetCompactBlocks, GetTransactions, TransIndex,
-    TransactionDigests,
+use crate::sync::{
+    message::{
+        GetBlockHashesByEpoch, GetBlockHeaderChain, GetBlockHeaders,
+        GetBlockTxn, GetBlocks, GetCompactBlocks, GetTransactions, TransIndex,
+    },
+    Error,
 };
+use cfx_types::H256;
 use metrics::{register_meter_with_group, Meter, MeterTimer};
 use network::{NetworkContext, PeerId};
 use parking_lot::{Mutex, RwLock};
@@ -465,18 +466,12 @@ impl RequestManager {
     }
 
     pub fn request_transactions(
-        &self, io: &NetworkContext, peer_id: PeerId,
-        transaction_digests: TransactionDigests,
+        &self, io: &NetworkContext, peer_id: PeerId, window_index: usize,
+        received_tx_ids: &Vec<TxPropagateId>,
     )
     {
         let _timer = MeterTimer::time_func(REQUEST_MANAGER_TX_TIMER.as_ref());
-
-        let window_index: usize = transaction_digests.window_index;
-        let (random_byte_vector, fixed_bytes_vector) =
-            transaction_digests.get_decomposed_short_ids();
-        let random_position: u8 = transaction_digests.random_position;
-
-        if fixed_bytes_vector.is_empty() {
+        if received_tx_ids.is_empty() {
             return;
         }
         let mut inflight_transactions =
@@ -490,24 +485,20 @@ impl RequestManager {
             let mut tx_ids = HashSet::new();
             let mut indices = Vec::new();
 
-            for i in 0..fixed_bytes_vector.len() {
-                if received_transactions.contains_txid(
-                    fixed_bytes_vector[i],
-                    random_byte_vector[i],
-                    random_position,
-                ) {
+            for (idx, tx_id) in received_tx_ids.iter().enumerate() {
+                if received_transactions.contains_txid(tx_id) {
                     // Already received
                     continue;
                 }
 
-                if !inflight_transactions.insert(fixed_bytes_vector[i]) {
+                if !inflight_transactions.insert(*tx_id) {
                     // Already being requested
                     continue;
                 }
 
-                let index = TransIndex::new((window_index, i));
+                let index = TransIndex::new((window_index, idx));
                 indices.push(index);
-                tx_ids.insert(fixed_bytes_vector[i]);
+                tx_ids.insert(*tx_id);
             }
 
             (indices, tx_ids)
