--- message/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "message"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-rlp = "0.3.0"
-cfx-types = { path = "../cfx_types" }
-cfx-bytes = { path = "../cfx_bytes" }
-primitives = { path = "../primitives" }
-log = "0.3"
